+++ pkg/pgmodel/ingestor/copier.go
@@ -145,6 +145,7 @@ func persistBatch(ctx context.Context, conn pgxconn.PgxConn, sw *seriesWriter, e
 	defer span.End()
 	batch := copyBatch(insertBatch)
 	err := sw.PopulateOrCreateSeries(ctx, batch)
+
 	if err != nil {
 		return fmt.Errorf("copier: writing series: %w", err)
 	}
@@ -209,6 +210,15 @@ func doInsertOrFallback(ctx context.Context, conn pgxconn.PgxConn, reqs ...copyR
 		if isPGUniqueViolation(err) {
 			err, _ = insertSeries(ctx, conn, true, reqs...)
 		}
+		if isEpochAbort(err) {
+			// Epoch abort is non-recoverable. Return error here instead of retrying.
+			log.Warn("msg", "Insert failed on epoch abort", "error", err)
+			for i := range reqs {
+				reqs[i].data.reportResults(NewInsertAbortError(err))
+				reqs[i].data.release()
+			}
+			return
+		}
 		if err != nil {
 			log.Error("msg", err)
 			insertBatchErrorFallback(ctx, conn, reqs...)
@@ -229,6 +239,17 @@ func isPGUniqueViolation(err error) bool {
 	}
 	pgErr, ok := err.(*pgconn.PgError)
 	if ok && pgErr.Code == "23505" {
+		return true
+	}
+	return false
+}
+
+func isEpochAbort(err error) bool {
+	if err == nil {
+		return false
+	}
+	pgErr, ok := err.(*pgconn.PgError)
+	if ok && pgErr.Code == "PS001" {
 		return true
 	}
 	return false
@@ -349,8 +370,8 @@ func insertSeries(ctx context.Context, conn pgxconn.PgxConn, onConflict bool, re
 	var sampleRows [][]interface{}
 	var exemplarRows [][]interface{}
 	insertStart := time.Now()
-	lowestEpoch := pgmodel.SeriesEpoch(math.MaxInt64)
 	lowestMinTime := int64(math.MaxInt64)
+	minCacheEpoch := pgmodel.SeriesEpoch(0)
 	tx, err := conn.BeginTx(ctx)
 	if err != nil {
 		return fmt.Errorf("failed to start transaction for inserting metrics: %v", err), lowestMinTime
@@ -365,6 +386,10 @@ func insertSeries(ctx context.Context, conn pgxconn.PgxConn, onConflict bool, re
 
 	for r := range reqs {
 		req := &reqs[r]
+		epoch := req.data.batch.SeriesCacheEpoch()
+		if minCacheEpoch == pgmodel.SeriesEpoch(0) || minCacheEpoch.After(epoch) {
+			minCacheEpoch = epoch
+		}
 		// Since seriesId order is not guaranteed we need to sort it to avoid row deadlock when duplicates are sent (eg. Prometheus retry)
 		// Samples inside series should be sorted by Prometheus
 		// We sort after PopulateOrCreateSeries call because we now have guarantees that all seriesIDs have been populated
@@ -409,10 +434,6 @@ func insertSeries(ctx context.Context, conn pgxconn.PgxConn, onConflict bool, re
 		)
 		if err != nil {
 			return err, lowestMinTime
-		}
-		epoch := visitor.LowestEpoch()
-		if epoch < lowestEpoch {
-			lowestEpoch = epoch
 		}
 		minTime := visitor.MinTime()
 		if minTime < lowestMinTime {
@@ -472,11 +493,12 @@ func insertSeries(ctx context.Context, conn pgxconn.PgxConn, onConflict bool, re
 		}
 	}
 
-	//note the epoch increment takes an access exclusive on the table before incrementing.
-	//thus we don't need row locking here. Note by doing this check at the end we can
-	//have some wasted work for the inserts before this fails but this is rare.
-	//avoiding an additional loop or memoization to find the lowest epoch ahead of time seems worth it.
-	row := tx.QueryRow(ctx, "SELECT CASE current_epoch > $1::BIGINT + 1 WHEN true THEN _prom_catalog.epoch_abort($1) END FROM _prom_catalog.ids_epoch LIMIT 1", int64(lowestEpoch))
+	// Note: The epoch increment takes an access exclusive on the table before
+	// incrementing, thus we don't need row locking here. By doing this check
+	// at the end we can have some wasted work for the inserts before this
+	// fails but this is rare. Avoiding an additional loop or memoization to
+	// find the lowest epoch ahead of time seems worth it.
+	row := tx.QueryRow(ctx, "SELECT CASE $1 <= delete_epoch WHEN true THEN _prom_catalog.epoch_abort($1) END FROM _prom_catalog.ids_epoch LIMIT 1", int64(minCacheEpoch))
 	var val []byte
 	if err = row.Scan(&val); err != nil {
 		return err, lowestMinTime