+++ pkg/pgmodel/ingestor/dispatcher.go
@@ -7,6 +7,7 @@ package ingestor
 import (
 	"context"
 	"fmt"
+	"github.com/pkg/errors"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -26,15 +27,16 @@ import (
 )
 
 const (
-	finalizeMetricCreation = "CALL _prom_catalog.finalize_metric_creation()"
-	getEpochSQL            = "SELECT current_epoch FROM _prom_catalog.ids_epoch LIMIT 1"
+	finalizeMetricCreation    = "CALL _prom_catalog.finalize_metric_creation()"
+	getEpochSQL               = "SELECT current_epoch, COALESCE(delete_epoch, 0) FROM _prom_catalog.ids_epoch LIMIT 1"
+	getStaleSeriesIDsArraySQL = "SELECT ARRAY_AGG(id) FROM _prom_catalog.series WHERE delete_epoch IS NOT NULL"
 )
 
 var ErrDispatcherClosed = fmt.Errorf("dispatcher is closed")
 
-// pgxDispatcher redirects incoming samples to the appropriate metricBatcher
+// PgxDispatcher redirects incoming samples to the appropriate metricBatcher
 // corresponding to the metric in the sample.
-type pgxDispatcher struct {
+type PgxDispatcher struct {
 	conn                   pgxconn.PgxConn
 	metricTableNames       cache.MetricCache
 	scache                 cache.SeriesCache
@@ -48,16 +50,50 @@ type pgxDispatcher struct {
 	doneChannel            chan struct{}
 	closed                 *uber_atomic.Bool
 	doneWG                 sync.WaitGroup
+	seriesRefreshLock      sync.Mutex
 }
 
-var _ model.Dispatcher = &pgxDispatcher{}
+type InsertAbortError struct {
+	err error
+}
 
-func newPgxDispatcher(conn pgxconn.PgxConn, mCache cache.MetricCache, scache cache.SeriesCache, eCache cache.PositionCache, cfg *Cfg) (*pgxDispatcher, error) {
+func NewInsertAbortError(err error) InsertAbortError {
+	return InsertAbortError{err}
+}
+
+func (e InsertAbortError) Error() string {
+	return e.err.Error()
+}
+
+var _ model.Dispatcher = &PgxDispatcher{}
+
+func newPgxDispatcher(conn pgxconn.PgxConn, mCache cache.MetricCache, scache cache.SeriesCache, eCache cache.PositionCache, cfg *Cfg) (*PgxDispatcher, error) {
 	numCopiers := cfg.NumCopiers
 	if numCopiers < 1 {
 		log.Warn("msg", "num copiers less than 1, setting to 1")
 		numCopiers = 1
 	}
+
+	var dbCurrentEpoch int64
+	// Bump the current epoch if it was still set to the initial value, and
+	// initialize the cache's epoch. Initializing the cache's epoch is crucial
+	// to ensure that ingestion will abort if a stale cache entry is present.
+	row := conn.QueryRow(context.Background(), "SELECT _prom_catalog.initialize_current_epoch(now())")
+	err := row.Scan(&dbCurrentEpoch)
+	if err != nil {
+		return nil, err
+	}
+	log.Info("msg", "Initializing epoch", "epoch", dbCurrentEpoch)
+	scache.SetCacheEpoch(model.SeriesEpoch(dbCurrentEpoch))
+
+	var epochDuration time.Duration
+	row = conn.QueryRow(context.Background(), "SELECT _prom_catalog.get_default_value('epoch_duration')::INTERVAL")
+	err = row.Scan(&epochDuration)
+	if err != nil {
+		return nil, err
+	}
+	var seriesEpochRefresh = epochDuration / 3 // This seems like a good enough rule of thumb
+	log.Info("msd", "Setting series epoch refresh", "series_epoch_refresh", seriesEpochRefresh)
 
 	// the copier read request channel retains the queue order between metrics
 	maxMetrics := 10000
@@ -80,14 +116,14 @@ func newPgxDispatcher(conn pgxconn.PgxConn, mCache cache.MetricCache, scache cac
 	if err != nil {
 		return nil, err
 	}
-	sw := NewSeriesWriter(conn, labelArrayOID, labelsCache)
+	sw := NewSeriesWriter(conn, labelArrayOID, labelsCache, scache)
 	elf := NewExamplarLabelFormatter(conn, eCache)
 
 	for i := 0; i < numCopiers; i++ {
 		go runCopier(conn, copierReadRequestCh, sw, elf)
 	}
 
-	inserter := &pgxDispatcher{
+	inserter := &PgxDispatcher{
 		conn:                   conn,
 		metricTableNames:       mCache,
 		scache:                 scache,
@@ -97,7 +133,7 @@ func newPgxDispatcher(conn pgxconn.PgxConn, mCache cache.MetricCache, scache cac
 		asyncAcks:              cfg.MetricsAsyncAcks,
 		copierReadRequestCh:    copierReadRequestCh,
 		// set to run at half our deletion interval
-		seriesEpochRefresh: time.NewTicker(30 * time.Minute),
+		seriesEpochRefresh: time.NewTicker(seriesEpochRefresh),
 		doneChannel:        make(chan struct{}),
 		closed:             uber_atomic.NewBool(false),
 	}
@@ -122,7 +158,7 @@ func newPgxDispatcher(conn pgxconn.PgxConn, mCache cache.MetricCache, scache cac
 	return inserter, nil
 }
 
-func (p *pgxDispatcher) runCompleteMetricCreationWorker() {
+func (p *PgxDispatcher) runCompleteMetricCreationWorker() {
 	for range p.completeMetricCreation {
 		err := p.CompleteMetricCreation(context.Background())
 		if err != nil {
@@ -131,56 +167,100 @@ func (p *pgxDispatcher) runCompleteMetricCreationWorker() {
 	}
 }
 
-func (p *pgxDispatcher) runSeriesEpochSync() {
-	epoch, err := p.refreshSeriesEpoch(model.InvalidSeriesEpoch)
-	// we don't have any great place to report errors, and if the
-	// connection recovers we can still make progress, so we'll just log it
-	// and continue execution
-	if err != nil {
-		log.Error("msg", "error refreshing the series cache", "err", err)
-	}
+func (p *PgxDispatcher) runSeriesEpochSync() {
+	p.RefreshSeriesEpoch()
 	for {
 		select {
 		case <-p.seriesEpochRefresh.C:
-			epoch, err = p.refreshSeriesEpoch(epoch)
-			if err != nil {
-				log.Error("msg", "error refreshing the series cache", "err", err)
-			}
+			p.RefreshSeriesEpoch()
 		case <-p.doneChannel:
 			return
 		}
 	}
 }
 
-func (p *pgxDispatcher) refreshSeriesEpoch(existingEpoch model.SeriesEpoch) (model.SeriesEpoch, error) {
-	dbEpoch, err := p.getServerEpoch()
+func (p *PgxDispatcher) getDatabaseEpochs() (model.SeriesEpoch, model.SeriesEpoch, error) {
+	var dbCurrentEpoch, dbDeleteEpoch int64
+	row := p.conn.QueryRow(context.Background(), getEpochSQL)
+	err := row.Scan(&dbCurrentEpoch, &dbDeleteEpoch)
 	if err != nil {
+		return 0, 0, err
+	}
+	return model.SeriesEpoch(dbCurrentEpoch), model.SeriesEpoch(dbDeleteEpoch), nil
+}
+
+func (p *PgxDispatcher) RefreshSeriesEpoch() {
+	// There can be multiple concurrent executions of `RefreshSeriesEpoch`. We
+	// use TryLock here to ensure that we don't have concurrent execution in
+	// this function, and that we don't have multiple back-to-back executions.
+	locked := p.seriesRefreshLock.TryLock()
+	if !locked {
+		log.Info("msg", "Skipping series cache refresh")
+		return
+	}
+	defer p.seriesRefreshLock.Unlock()
+	log.Info("msg", "Refreshing series cache epoch")
+	cacheCurrentEpoch := p.scache.CacheEpoch()
+	dbCurrentEpoch, dbDeleteEpoch, err := p.getDatabaseEpochs()
+	log.Info("msg", "RefreshSeriesEpoch", "cacheCurrentEpoch", cacheCurrentEpoch, "dbCurrentEpoch", dbCurrentEpoch, "dbDeleteEpoch", dbDeleteEpoch)
+	if err != nil {
+		log.Warn("msg", "Unable to get database epoch data, will reset series and inverted labels caches", "err", err.Error())
 		// Trash the cache just in case an epoch change occurred, seems safer
-		p.scache.Reset()
+		p.scache.Reset(dbCurrentEpoch)
 		// Also trash the inverted labels cache, which can also be invalidated when the series cache is
 		p.invertedLabelsCache.Reset()
-		return model.InvalidSeriesEpoch, err
+		return
 	}
-	if existingEpoch == model.InvalidSeriesEpoch || dbEpoch != existingEpoch {
-		p.scache.Reset()
-		// If the series cache needs to be invalidated, so does the inverted labels cache
+	if cacheCurrentEpoch.After(dbCurrentEpoch) {
+		log.Warn("msg", "The connector's cache epoch is greater than the databases. This is unexpected", "connector_current_epoch", cacheCurrentEpoch, "database_current_epoch", dbCurrentEpoch)
+		// Reset the caches. It's not really clear what would lead to this
+		// situation, but if it does then something must have gone wrong...
+		p.scache.Reset(dbCurrentEpoch)
+		p.invertedLabelsCache.Reset()
+		return
+	}
+	if dbDeleteEpoch.AfterEq(cacheCurrentEpoch) {
+		// The current cache epoch has been overtaken by the database'
+		// delete_epoch.
+		// The only way to recover from this situation is to reset our caches
+		// and let them repopulate.
+		log.Warn("msg", "Cache epoch was overtaken by the database's delete epoch, resetting series and labels caches", "cache_epoch", cacheCurrentEpoch, "delete_epoch", dbDeleteEpoch)
+		p.scache.Reset(dbCurrentEpoch)
+		p.invertedLabelsCache.Reset()
+		return
+	} else {
+		start := time.Now()
+		staleSeriesIds, err := GetStaleSeriesIDs(p.conn)
+		if err != nil {
+			log.Warn("msg", "Error getting series ids, unable to update series cache", "err", err.Error())
+			return
+		}
+		log.Info("msg", "Epoch change noticed, fetched stale series from db", "count", len(staleSeriesIds), "duration", time.Since(start))
+		start = time.Now()
+		evictCount := p.scache.EvictSeriesById(staleSeriesIds, dbCurrentEpoch)
+		log.Info("msg", "Removed stale series", "count", evictCount, "duration", time.Since(start)) // Before merging in master, change the level to Debug.
+		// Trash the inverted labels cache. Technically, we could determine
+		// which label entries need to be removed from the cache, as we have
+		// done for the series cache, but that doesn't seem to really be worth
+		// it.
 		p.invertedLabelsCache.Reset()
 	}
-	return dbEpoch, nil
 }
 
-func (p *pgxDispatcher) getServerEpoch() (model.SeriesEpoch, error) {
-	var newEpoch int64
-	row := p.conn.QueryRow(context.Background(), getEpochSQL)
-	err := row.Scan(&newEpoch)
+func GetStaleSeriesIDs(conn pgxconn.PgxConn) ([]model.SeriesID, error) {
+	staleSeriesIDs := make([]int64, 0)
+	err := conn.QueryRow(context.Background(), getStaleSeriesIDsArraySQL).Scan(&staleSeriesIDs)
 	if err != nil {
-		return -1, err
+		return nil, fmt.Errorf("error getting stale series ids from db: %w", err)
 	}
-
-	return model.SeriesEpoch(newEpoch), nil
+	staleSeriesIdsAsSeriesId := make([]model.SeriesID, len(staleSeriesIDs))
+	for i := range staleSeriesIDs {
+		staleSeriesIdsAsSeriesId[i] = model.SeriesID(staleSeriesIDs[i])
+	}
+	return staleSeriesIdsAsSeriesId, nil
 }
 
-func (p *pgxDispatcher) CompleteMetricCreation(ctx context.Context) error {
+func (p *PgxDispatcher) CompleteMetricCreation(ctx context.Context) error {
 	if p.closed.Load() {
 		return ErrDispatcherClosed
 	}
@@ -193,7 +273,7 @@ func (p *pgxDispatcher) CompleteMetricCreation(ctx context.Context) error {
 	return err
 }
 
-func (p *pgxDispatcher) Close() {
+func (p *PgxDispatcher) Close() {
 	if p.closed.Load() {
 		return
 	}
@@ -215,7 +295,7 @@ func (p *pgxDispatcher) Close() {
 // actually inserted) and any error.
 // Though we may insert data to multiple tables concurrently, if asyncAcks is
 // unset this function will wait until _all_ the insert attempts have completed.
-func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64, error) {
+func (p *PgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64, error) {
 	if p.closed.Load() {
 		return 0, ErrDispatcherClosed
 	}
@@ -225,6 +305,7 @@ func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64
 		numRows      uint64
 		maxt         int64
 		rows         = dataTS.Rows
+		seriesEpoch  = dataTS.SeriesCacheEpoch
 		workFinished = new(sync.WaitGroup)
 	)
 	workFinished.Add(len(rows))
@@ -240,7 +321,7 @@ func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64
 				maxt = ts
 			}
 		}
-		p.getMetricBatcher(metricName) <- &insertDataRequest{spanCtx: span.SpanContext(), metric: metricName, data: data, finished: workFinished, errChan: errChan}
+		p.getMetricBatcher(metricName) <- &insertDataRequest{spanCtx: span.SpanContext(), metric: metricName, seriesCacheEpoch: seriesEpoch, data: data, finished: workFinished, errChan: errChan}
 	}
 	span.SetAttributes(attribute.Int64("num_rows", int64(numRows)))
 	span.SetAttributes(attribute.Int("num_metrics", len(rows)))
@@ -258,6 +339,9 @@ func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64
 		case err = <-errChan:
 		default:
 		}
+		if errors.As(err, &InsertAbortError{}) {
+			p.RefreshSeriesEpoch()
+		}
 		reportMetricsTelemetry(maxt, numRows, 0)
 		close(errChan)
 	} else {
@@ -270,6 +354,9 @@ func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64
 			}
 			close(errChan)
 			if err != nil {
+				if errors.As(err, &InsertAbortError{}) {
+					p.RefreshSeriesEpoch()
+				}
 				log.Error("msg", fmt.Sprintf("error on async send, dropping %d datapoints", numRows), "err", err)
 			}
 			reportMetricsTelemetry(maxt, numRows, 0)
@@ -279,7 +366,7 @@ func (p *pgxDispatcher) InsertTs(ctx context.Context, dataTS model.Data) (uint64
 	return numRows, err
 }
 
-func (p *pgxDispatcher) InsertMetadata(ctx context.Context, metadata []model.Metadata) (uint64, error) {
+func (p *PgxDispatcher) InsertMetadata(ctx context.Context, metadata []model.Metadata) (uint64, error) {
 	if p.closed.Load() {
 		return 0, ErrDispatcherClosed
 	}
@@ -309,7 +396,7 @@ func reportMetricsTelemetry(maxTs int64, numSamples, numMetadata uint64) {
 }
 
 // Get the handler for a given metric name, creating a new one if none exists
-func (p *pgxDispatcher) getMetricBatcher(metric string) chan<- *insertDataRequest {
+func (p *PgxDispatcher) getMetricBatcher(metric string) chan<- *insertDataRequest {
 	batcher, ok := p.batchers.Load(metric)
 	if !ok {
 		// The ordering is important here: we need to ensure that every call
@@ -330,11 +417,12 @@ func (p *pgxDispatcher) getMetricBatcher(metric string) chan<- *insertDataReques
 }
 
 type insertDataRequest struct {
-	spanCtx  trace.SpanContext
-	metric   string
-	finished *sync.WaitGroup
-	data     []model.Insertable
-	errChan  chan error
+	spanCtx          trace.SpanContext
+	metric           string
+	seriesCacheEpoch model.SeriesEpoch
+	finished         *sync.WaitGroup
+	data             []model.Insertable
+	errChan          chan error
 }
 
 func (idr *insertDataRequest) reportResult(err error) {