+++ pkg/pgmodel/ingestor/series_writer.go
@@ -7,7 +7,6 @@ package ingestor
 import (
 	"context"
 	"fmt"
-
 	"github.com/jackc/pgtype"
 	"github.com/timescale/promscale/pkg/log"
 	"github.com/timescale/promscale/pkg/pgmodel/cache"
@@ -27,6 +26,7 @@ type seriesWriter struct {
 	conn          pgxconn.PgxConn
 	labelArrayOID uint32
 	labelsCache   *cache.InvertedLabelsCache
+	seriesCache   cache.SeriesCache
 }
 
 type SeriesVisitor interface {
@@ -35,8 +35,8 @@ type SeriesVisitor interface {
 
 func labelArrayTranscoder() pgtype.ValueTranscoder { return &pgtype.Int4Array{} }
 
-func NewSeriesWriter(conn pgxconn.PgxConn, labelArrayOID uint32, labelsCache *cache.InvertedLabelsCache) *seriesWriter {
-	return &seriesWriter{conn, labelArrayOID, labelsCache}
+func NewSeriesWriter(conn pgxconn.PgxConn, labelArrayOID uint32, labelsCache *cache.InvertedLabelsCache, seriesCache cache.SeriesCache) *seriesWriter {
+	return &seriesWriter{conn, labelArrayOID, labelsCache, seriesCache}
 }
 
 type perMetricInfo struct {
@@ -59,6 +59,7 @@ func (h *seriesWriter) PopulateOrCreateSeries(ctx context.Context, sv SeriesVisi
 	defer span.End()
 	infos := make(map[string]*perMetricInfo)
 	seriesCount := 0
+
 	err := sv.VisitSeries(func(metricInfo *pgmodel.MetricInfo, series *model.Series) error {
 		if !series.IsSeriesIDSet() {
 			metricName := series.MetricName()
@@ -124,7 +125,7 @@ func (h *seriesWriter) PopulateOrCreateSeries(ctx context.Context, sv SeriesVisi
 	//the labels for multiple series in same txn as we are creating the series,
 	//the ordering of label creation can only be canonical within a series and
 	//not across series.
-	dbEpoch, err := h.fillLabelIDs(ctx, infos, labelMap)
+	err = h.fillLabelIDs(ctx, infos, labelMap)
 	if err != nil {
 		return fmt.Errorf("error setting series ids: %w", err)
 	}
@@ -142,7 +143,7 @@ func (h *seriesWriter) PopulateOrCreateSeries(ctx context.Context, sv SeriesVisi
 			continue
 		}
 
-		//transaction per metric to avoid cross-metric locks
+		// transaction per metric to avoid cross-metric locks
 		batch.Queue("BEGIN;")
 		batch.Queue(seriesInsertSQL, info.metricInfo.MetricID, info.metricInfo.TableName, info.labelArraySet)
 		batch.Queue("COMMIT;")
@@ -176,7 +177,7 @@ func (h *seriesWriter) PopulateOrCreateSeries(ctx context.Context, sv SeriesVisi
 			if err != nil {
 				return fmt.Errorf("error setting series_id: cannot scan series_id: %w", err)
 			}
-			info.series[int(ordinality)-1].SetSeriesID(id, dbEpoch)
+			info.series[int(ordinality)-1].SetSeriesID(id)
 			count++
 		}
 		if err := res.Err(); err != nil {
@@ -196,7 +197,7 @@ func (h *seriesWriter) PopulateOrCreateSeries(ctx context.Context, sv SeriesVisi
 	return nil
 }
 
-func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMetricInfo, labelMap map[cache.LabelKey]cache.LabelInfo) (model.SeriesEpoch, error) {
+func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMetricInfo, labelMap map[cache.LabelKey]cache.LabelInfo) error {
 	_, span := tracer.Default().Start(ctx, "fill-label-ids")
 	defer span.End()
 	//we cannot use the label cache here because that maps label ids => name, value.
@@ -204,13 +205,6 @@ func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMe
 	//we may want a new cache for that, at a later time.
 
 	batch := h.conn.NewBatch()
-	var dbEpoch model.SeriesEpoch
-
-	// The epoch will never decrease, so we can check it once at the beginning,
-	// at worst we'll store too small an epoch, which is always safe
-	batch.Queue("BEGIN;")
-	batch.Queue(getEpochSQL)
-	batch.Queue("COMMIT;")
 
 	infoBatches := make([]*perMetricInfo, 0)
 	items := 0
@@ -226,11 +220,11 @@ func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMe
 			}
 			namesSlice, err := names.Slice(i, high)
 			if err != nil {
-				return dbEpoch, fmt.Errorf("error filling labels: slicing names: %w", err)
+				return fmt.Errorf("error filling labels: slicing names: %w", err)
 			}
 			valuesSlice, err := values.Slice(i, high)
 			if err != nil {
-				return dbEpoch, fmt.Errorf("error filling labels: slicing values: %w", err)
+				return fmt.Errorf("error filling labels: slicing values: %w", err)
 			}
 			batch.Queue("BEGIN;")
 			batch.Queue("SELECT * FROM _prom_catalog.get_or_create_label_ids($1, $2, $3, $4)", metricName, info.metricInfo.TableName, namesSlice, valuesSlice)
@@ -250,25 +244,14 @@ func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMe
 
 	br, err := h.conn.SendBatch(context.Background(), batch)
 	if err != nil {
-		return dbEpoch, fmt.Errorf("error filling labels: %w", err)
+		return fmt.Errorf("error filling labels: %w", err)
 	}
 	defer br.Close()
-
-	if _, err := br.Exec(); err != nil {
-		return dbEpoch, fmt.Errorf("error filling labels on begin: %w", err)
-	}
-	err = br.QueryRow().Scan(&dbEpoch)
-	if err != nil {
-		return dbEpoch, fmt.Errorf("error filling labels: %w", err)
-	}
-	if _, err := br.Exec(); err != nil {
-		return dbEpoch, fmt.Errorf("error filling labels on commit: %w", err)
-	}
 
 	var count int
 	for _, info := range infoBatches {
 		if _, err := br.Exec(); err != nil {
-			return dbEpoch, fmt.Errorf("error filling labels on begin label batch: %w", err)
+			return fmt.Errorf("error filling labels on begin label batch: %w", err)
 		}
 
 		err := func() error {
@@ -306,16 +289,16 @@ func (h *seriesWriter) fillLabelIDs(ctx context.Context, infos map[string]*perMe
 			return nil
 		}()
 		if err != nil {
-			return dbEpoch, err
+			return err
 		}
 		if _, err := br.Exec(); err != nil {
-			return dbEpoch, fmt.Errorf("error filling labels on commit label batch: %w", err)
+			return fmt.Errorf("error filling labels on commit label batch: %w", err)
 		}
 	}
 	if count != items {
-		return dbEpoch, fmt.Errorf("error filling labels: not filling as many items as expected: %v vs %v", count, items)
+		return fmt.Errorf("error filling labels: not filling as many items as expected: %v vs %v", count, items)
 	}
-	return dbEpoch, nil
+	return nil
 }
 
 func (h *seriesWriter) buildLabelArrays(ctx context.Context, infos map[string]*perMetricInfo, labelMap map[cache.LabelKey]cache.LabelInfo) error {