+++ CHANGELOG.md
@@ -23,6 +23,7 @@ We use the following categories for changes:
   configuration. Supersedes `startup.dataset.config` which accepts a string
   instead of a mapping [#1737]
 - Add alert to notify about duplicate sample/metric ingestion. [#1688]
+- Implemented proper invalidation of the series cache [#1752]
 
 ### Changed
 - Reduced the verbosity of the logs emitted by the vacuum engine [#1715]